Refactor ArticleService with different service
