+++ src/main/kotlin/io/github/raeperd/realworldspringbootkotlin/domain/ArticleService.kt
@@ -55,54 +55,6 @@ class ArticleService(
             ?: articles.map { it.toArticleDTO() }
     }
 
-    fun favoriteArticle(userId: Long, slug: String): ArticleDTO {
-        val user = userRepository.findUserByIdOrThrow(userId)
-        val article = articleRepository.findArticleBySlugOrThrow(slug)
-        user.favoriteArticle(article)
-        return articleRepository.saveArticle(article)
-            .toArticleDTO(user)
-    }
-
-    fun unfavoriteArticle(userId: Long, slug: String): ArticleDTO {
-        val user = userRepository.findUserByIdOrThrow(userId)
-        val article = articleRepository.findArticleBySlugOrThrow(slug)
-        if (!article.isFavoritedByUser(user)) {
-            return article.toArticleDTO()
-        }
-        user.unfavoriteArticle(article)
-        return articleRepository.saveArticle(article)
-            .toArticleDTO(user)
-    }
-
-    private fun Article.toArticleDTO(user: User): ArticleDTO {
-        return ArticleDTO(
-            slug = slug,
-            title = title,
-            description = description,
-            body = body,
-            tagList = tagList.map { it.toString() },
-            author = author.toProfileDTO(),
-            createdAt = createdAt,
-            updatedAt = updatedAt,
-            favoritesCount = favoritesCount,
-            favorited = isFavoritedByUser(user)
-        )
-    }
-
-    private fun Article.toArticleDTO(): ArticleDTO {
-        return ArticleDTO(
-            slug = slug,
-            title = title,
-            description = description,
-            body = body,
-            tagList = tagList.map { it.toString() },
-            author = author.toProfileDTO(),
-            createdAt = createdAt,
-            updatedAt = updatedAt,
-            favoritesCount = favoritesCount,
-            favorited = false
-        )
-    }
 }
 
 data class ArticleDTO(
@@ -117,3 +69,33 @@ data class ArticleDTO(
     val favoritesCount: Int,
     val favorited: Boolean
 )
+
+fun Article.toArticleDTO(user: User): ArticleDTO {
+    return ArticleDTO(
+        slug = slug,
+        title = title,
+        description = description,
+        body = body,
+        tagList = tagList.map { it.toString() },
+        author = author.toProfileDTO(),
+        createdAt = createdAt,
+        updatedAt = updatedAt,
+        favoritesCount = favoritesCount,
+        favorited = isFavoritedByUser(user)
+    )
+}
+
+fun Article.toArticleDTO(): ArticleDTO {
+    return ArticleDTO(
+        slug = slug,
+        title = title,
+        description = description,
+        body = body,
+        tagList = tagList.map { it.toString() },
+        author = author.toProfileDTO(),
+        createdAt = createdAt,
+        updatedAt = updatedAt,
+        favoritesCount = favoritesCount,
+        favorited = false
+    )
+}
