+++ src/main/kotlin/io/github/raeperd/realworldspringbootkotlin/web/ArticleFavoriteRestController.kt
@@ -1,6 +1,6 @@
 package io.github.raeperd.realworldspringbootkotlin.web
 
-import io.github.raeperd.realworldspringbootkotlin.domain.ArticleService
+import io.github.raeperd.realworldspringbootkotlin.domain.ArticleFavoriteService
 import io.github.raeperd.realworldspringbootkotlin.domain.JWTPayload
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -9,15 +9,15 @@ import org.springframework.web.bind.annotation.RestController
 
 @RestController
 class ArticleFavoriteRestController(
-    private val articleService: ArticleService
+    private val articleFavoriteService: ArticleFavoriteService
 ) {
     @PostMapping("/articles/{slug}/favorite")
     fun postArticleFavoriteBySlug(@PathVariable slug: String, payload: JWTPayload): ArticleModel {
-        return articleService.favoriteArticle(payload.sub, slug).toArticleModel()
+        return articleFavoriteService.favoriteArticle(payload.sub, slug).toArticleModel()
     }
 
     @DeleteMapping("/articles/{slug}/favorite")
     fun deleteArticleFavoriteBySlug(@PathVariable slug: String, payload: JWTPayload): ArticleModel {
-        return articleService.unfavoriteArticle(payload.sub, slug).toArticleModel()
+        return articleFavoriteService.unfavoriteArticle(payload.sub, slug).toArticleModel()
     }
 }
